+++ Sming/Arch/Esp8266/Components/spi_flash/flashmem.c
@@ -16,7 +16,6 @@
  ****/
 
 #include "include/esp_spi_flash.h"
-#include "espinc/peri.h"
 
 extern char _flash_code_end[];
